Move TestCallback & TestEventCallback out of messages package.
