+++ formula/src/test/java/com/instacart/formula/FormulaRuntimeTest.kt
@@ -2,8 +2,8 @@ package com.instacart.formula
 
 import com.google.common.truth.Truth.assertThat
 import com.instacart.formula.streams.EmptyStream
-import com.instacart.formula.test.messages.TestCallback
-import com.instacart.formula.test.messages.TestEventCallback
+import com.instacart.formula.test.TestCallback
+import com.instacart.formula.test.TestEventCallback
 import com.instacart.formula.test.test
 import io.reactivex.Observable
 import org.junit.Test
